Extend private registry support to balena build and deploy commands
Resolves: #1116
Change-type: minor
